+++ pubspec.yaml
@@ -20,7 +20,7 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   intl: ^0.17.0
-
+  http: ^0.13.4
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
